Used BigInteger instead of Integer for version subpart in case the file name contains a very long string of digits
git-svn-id: https://svn.code.sf.net/p/sweethome3d/code/trunk/SweetHome3D@4073 03e323e1-6c6c-46a5-9f71-c4ce5d8c6204

+++ src/com/eteks/sweethome3d/tools/OperatingSystem.java
@@ -22,6 +22,7 @@ package com.eteks.sweethome3d.tools;
 import java.io.File;
 import java.io.FileFilter;
 import java.io.IOException;
+import java.math.BigInteger;
 import java.security.AccessControlException;
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -201,7 +202,7 @@ public class OperatingSystem {
           for ( ; i < part.length() && Character.isDigit(c = part.charAt(i)); i++) {
             subPart.append(c);
           }
-          versionParts.add(new Integer(subPart.toString()));
+          versionParts.add(new BigInteger(subPart.toString()));
         } else {
           for ( ; i < part.length() && !Character.isDigit(c = part.charAt(i)); i++) {
             subPart.append(c);
